+++ packages/core/src/data/strategies/__tests__/PostOrPostsFetchStrategy.ts
@@ -183,7 +183,7 @@ describe('PostOrPostsFetchStrategy', () => {
 		);
 
 		await expect(() => fetchStrategy.fetcher('', params)).rejects.toThrow(
-			'Neither single or archive returned data: The request to /wp-json/wp/v2/posts?_embed=true&categories=distinctio-rerum-ratione-maxime-repudiandae-laboriosam-quam returned no data, Post was found but did not match current path: "/uncategorized/distinctio-rerum-ratione-maxime-repudiandae-laboriosam-quam"',
+			'Neither single or archive returned data: The request to /wp-json/wp/v2/posts?_embed=true&categories=distinctio-rerum-ratione-maxime-repudiandae-laboriosam-quam returned no data, Post #54 - "https://js1.10up.com/2020/05/07/distinctio-rerum-ratione-maxime-repudiandae-laboriosam-quam/" was found but did not match current path: "/uncategorized/distinctio-rerum-ratione-maxime-repudiandae-laboriosam-quam"',
 		);
 
 		// now make it work by faking full path
@@ -243,7 +243,7 @@ describe('PostOrPostsFetchStrategy', () => {
 			useWordPressPlugin: true,
 		});
 
-		const params = fetchStrategy.getParamsFromURL('/page/2');
+		const params = fetchStrategy.getParamsFromURL('/page/1');
 		const response = await fetchStrategy.fetcher(
 			'',
 			merge(params, { priority: 'single', routeMatchStrategy: 'archive' }),
@@ -257,7 +257,7 @@ describe('PostOrPostsFetchStrategy', () => {
 			useWordPressPlugin: true,
 		});
 
-		let params = fetchStrategy.getParamsFromURL('/page/2');
+		let params = fetchStrategy.getParamsFromURL('/page/1');
 		let response = await fetchStrategy.fetcher(
 			'',
 			merge(params, { priority: 'single', routeMatchStrategy: 'single' }),
@@ -265,7 +265,7 @@ describe('PostOrPostsFetchStrategy', () => {
 		expect(response.result.isSingle).toBeFalsy();
 		expect(response.result.isArchive).toBeTruthy();
 
-		params = fetchStrategy.getParamsFromURL('/page/2');
+		params = fetchStrategy.getParamsFromURL('/page/1');
 		response = await fetchStrategy.fetcher(
 			'',
 			merge(params, { priority: 'archive', routeMatchStrategy: 'single' }),