+++ packages/core/src/react/hooks/__tests__/useFetchPost.tsx
@@ -1,14 +1,21 @@
 import { renderHook, waitFor } from '@testing-library/react';
 import * as React from 'react';
 import { expectTypeOf } from 'expect-type';
+import { SWRConfig } from 'swr';
 import { DRAFT_POST_ID, VALID_AUTH_TOKEN } from '../../../../test/server';
 import { PostEntity, PostParams } from '../../../data';
 import { SettingsProvider } from '../../provider';
 import { useFetchPost } from '../useFetchPost';
+import * as useFetchModule from '../useFetch';
+import { mockUseFetchErrorResponse } from '../mocks';
 
 describe('useFetchPost', () => {
 	const wrapper = ({ children }) => {
-		return <SettingsProvider settings={{ sourceUrl: '' }}>{children}</SettingsProvider>;
+		return (
+			<SWRConfig value={{ provider: () => new Map() }}>
+				<SettingsProvider settings={{ sourceUrl: '' }}>{children}</SettingsProvider>
+			</SWRConfig>
+		);
 	};
 
 	it('throws errors if accessing data before fetch', async () => {
@@ -41,6 +48,31 @@ describe('useFetchPost', () => {
 				'modi-qui-dignissimos-sed-assumenda-sint-iusto',
 			),
 		);
+	});
+
+	it('handles response if has error or there is no data', async () => {
+		const spyUseFetch = jest
+			.spyOn(useFetchModule, 'useFetch')
+			.mockReturnValueOnce(mockUseFetchErrorResponse);
+		const { result } = renderHook(() => useFetchPost({}), {
+			wrapper,
+		});
+
+		const expectedKeys = ['error', 'loading', 'data', 'isMainQuery'];
+		const returnedKeys = Object.keys(result.current);
+		const missingKeys = returnedKeys.filter((key) => !expectedKeys.includes(key));
+
+		await waitFor(() => {
+			expect(missingKeys).toHaveLength(0);
+			expect(spyUseFetch).toHaveBeenCalledTimes(1);
+			expect(result.current.error).toBe('Not found');
+			expect(result.current.loading).toBe(false);
+			expect(() => result.current.data).not.toThrow();
+			expect(() => result.current.data?.post.title).toThrow();
+			expect(result.current.isMainQuery).toBe(true);
+		});
+
+		spyUseFetch.mockRestore();
 	});
 
 	it('fetch by id', async () => {
@@ -158,7 +190,15 @@ describe('useFetchPost', () => {
 
 	it('reads param from the url and sets isMainQuery flag', async () => {
 		const { result } = renderHook(
-			() => useFetchPost({}, {}, '/modi-qui-dignissimos-sed-assumenda-sint-iusto'),
+			() =>
+				useFetchPost(
+					{
+						fullPath:
+							'https://js1.10up.com/2020/05/07/modi-qui-dignissimos-sed-assumenda-sint-iusto/',
+					},
+					{},
+					'/modi-qui-dignissimos-sed-assumenda-sint-iusto/',
+				),
 			{
 				wrapper,
 			},
@@ -173,7 +213,7 @@ describe('useFetchPost', () => {
 		});
 
 		const { result: secondResult } = renderHook(
-			() => useFetchPost({ slug: 'modi-qui-dignissimos-sed-assumenda-sint-iusto' }),
+			() => useFetchPost({ slug: 'modi-qui-dignissimos-sed-assumenda-sint-iusto' }, {}),
 			{
 				wrapper,
 			},